refactoring

+++ utils.py
@@ -12,7 +12,7 @@ def get_mask_from_lengths(lengths):
 
 def load_wav_to_torch(full_path):
     sampling_rate, data = read(full_path)
-    return torch.FloatTensor(data.astype(np.float32)), sampling_rate
+    return torch.FloatTensor(data), sampling_rate
 
 
 def load_filepaths_and_text(filename, split="|"):
